refact: sticker 관련 로직 위치 변경
