+++ DonDog-iOS/DonDog-iOS/Presentation/Features/Post/ViewModels/StickerViewModel.swift
@@ -1,13 +1,15 @@
 //
-//  StickerSheetViewModel.swift
+//  StickerViewModel.swift
 //  DonDog-iOS
 //
 //  Created by 이서현 on 11/7/25.
 //
 
 import Combine
+import CoreGraphics
 import FirebaseAuth
 import FirebaseFirestore
+import Foundation
 import SwiftUI
 
 struct AttachedSticker: Identifiable, Codable {
@@ -18,7 +20,7 @@ struct AttachedSticker: Identifiable, Codable {
     var rotation: Angle
 }
 
-final class StickerSheetViewModel: ObservableObject {
+final class StickerViewModel: ObservableObject {
     @Published var selectedCategory: StickerCategory = .affection
     @Published var itemsByCategory: [StickerCategory: [StickerItem]] = StickerCategoryData.itemsByCategory
     
@@ -27,6 +29,9 @@ final class StickerSheetViewModel: ObservableObject {
     
     private let dataManager: DataManagerProtocol = DataManager.shared
     private let connectUserInfo = UserPairingStore.shared
+    
+    @Published var stickers: [AttachedSticker] = []
+    @Published var selectedStickerID: UUID?
     
     func returnStickerItems(for category: StickerCategory) -> [StickerItem] {
         itemsByCategory[category] ?? []
@@ -53,5 +58,45 @@ final class StickerSheetViewModel: ObservableObject {
             print("스티커 이미지 로드 실패: \(error)")
         }
         loadingItemIDs.remove(item.id)
+    }
+    
+    func addSticker(named imageName: String) {
+        let newSticker = AttachedSticker(
+            imageName: imageName,
+            position: .zero,
+            scale: 1.0,
+            rotation: .zero
+        )
+        stickers.append(newSticker)
+        selectedStickerID = newSticker.id
+    }
+    
+    func removeSticker(_ sticker: AttachedSticker) {
+        stickers.removeAll { $0.id == sticker.id }
+    }
+    
+    func saveStickers() {
+        // print()로 상태 출력
+        print("===== StickerData 상태 =====")
+        for s in stickers {
+            print("ID: \(s.id)")
+            print("Image: \(s.imageName)")
+            print("Position: \(s.position)")
+            print("Scale: \(s.scale)")
+            print("Rotation: \(s.rotation.degrees)°")
+            print("---------------------------")
+        }
+        print("============================\n")
+        
+        if let encoded = try? JSONEncoder().encode(stickers) {
+            UserDefaults.standard.set(encoded, forKey: "savedStickers")
+        }
+    }
+    
+    func loadSavedStickers() {
+        if let data = UserDefaults.standard.data(forKey: "savedStickers"),
+            let decoded = try? JSONDecoder().decode([AttachedSticker].self, from: data) {
+            stickers = decoded
+        }
     }
 }
